adjust minting of shares, refunding and e2e test
